+++ src/libs/protocols/turbos/index.ts
@@ -3,14 +3,9 @@ import { Transaction } from "@mysten/sui/transactions";
 import { SUI_CLOCK_OBJECT_ID } from "@mysten/sui/utils";
 import { BaseContract } from "../base";
 import { getDefaultSqrtPriceLimit } from "../utils";
-import { ONE_MINUTE } from "./constants";
 import { SuiUtils } from "../../../utils/sui";
 
-const PACKAGE_ID =
-  "0x1a3c42ded7b75cdf4ebc7c7b7da9d1e1db49f16fcdca934fac003f35f39ecad9";
-const MODULE_NAME = "swap_router";
-const VERSION =
-  "0xf1cf0e81048df168ebeb1b8030fad24b3e0b53ae827c25053fff0779c1445b6f";
+const ONE_MINUTE = 60 * 1000;
 
 export class TurbosContract extends BaseContract {
   async swap(tx: Transaction) {
@@ -20,18 +15,17 @@ export class TurbosContract extends BaseContract {
       address: this.currentAccount,
     };
 
-    const typeArguments = this.getTypeParams();
-
+    const { package: PACKAGE_ID, version: VERSION } = this.config.turbos;
     const inputAmount = this.getInputCoinValue(tx);
     const [tokenOut, tokenIn] = tx.moveCall({
-      target: `${PACKAGE_ID}::${MODULE_NAME}::swap_${
+      target: `${PACKAGE_ID}::swap_router::swap_${
         a2b ? "a_b" : "b_a"
       }_with_return_`,
-      typeArguments: typeArguments,
+      typeArguments: this.getTypeParams(),
       arguments: [
         tx.object(poolId),
         tx.makeMoveVec({
-          elements: [this.inputCoinObject as any],
+          elements: [this.inputCoinObject],
         }),
         inputAmount,
         tx.pure.u64(0),